update to match upstream changes
