+++ src/ops.rs
@@ -1,3 +1,5 @@
 pub mod create;
 pub mod fetch;
+pub mod graph;
+pub mod info;
 pub mod update;
